New player page: code quality overhaul, fix colors/icons, use new client package
