+++ components/player/PlayerCard.tsx
@@ -5,7 +5,11 @@ import { ExternalLink, User } from 'lucide-react';
 import PlayerRulesetSelector from '../buttons/RulesetSelector';
 import { Card } from '../ui/card';
 
-export default function PlayerCard({ player }: { player: PlayerCompactDTO }) {
+interface PlayerCardProps {
+  player: PlayerCompactDTO;
+}
+
+export default function PlayerCard({ player }: PlayerCardProps) {
   return (
     <Card className="flex flex-row flex-wrap justify-between border-none bg-popover p-4">
       <div className="flex min-w-[250px] flex-1 items-center gap-3 rounded-lg">
