+++ package.json
@@ -15,7 +15,7 @@
   "dependencies": {
     "@alduino/humanizer": "^1.1.0",
     "@hookform/resolvers": "^4.1.3",
-    "@osu-tournament-rating/otr-api-client": "^0.0.32-dev.2",
+    "@osu-tournament-rating/otr-api-client": "^0.0.34",
     "@radix-ui/react-alert-dialog": "^1.1.6",
     "@radix-ui/react-avatar": "^1.1.3",
     "@radix-ui/react-checkbox": "^1.1.5",
